refactor: level resource
